+++ packages/2d/rollup.config.mjs
@@ -1,0 +1,16 @@
+import makeConfig from '@motion-canvas/internal/rollup/makeConfig.mjs';
+
+export default makeConfig({
+  plugins: [
+    {
+      resolveId(id) {
+        if (id.startsWith('@motion-canvas/core')) {
+          return {
+            id: '@motion-canvas/core',
+            external: true,
+          };
+        }
+      },
+    },
+  ],
+});