+++ packages/core/src/plugin/DefaultPlugin.ts
@@ -3,6 +3,8 @@ import {ImageExporter} from '../app';
 
 /**
  * The default plugin included in every Motion Canvas project.
+ *
+ * @internal
  */
 export default makePlugin({
   name: 'mc-default-plugin',