+++ packages/docs/docusaurus.config.js
@@ -3,7 +3,8 @@
 /** @type {import('@docusaurus/types').Config} */
 const config = {
   title: 'Motion Canvas',
-  tagline: 'Visualize complex ideas programmatically',
+  tagline:
+    'A TypeScript library for creating animated videos using the Canvas API.',
   url: 'https://motioncanvas.io',
   baseUrl: '/',
   onBrokenLinks: 'throw',
@@ -121,6 +122,7 @@ const config = {
   ],
   plugins: [
     './typedoc',
+    './bundle',
     [
       '@docusaurus/plugin-content-docs',
       {