+++ packages/docs/src/components/Fiddle/SharedPlayer.ts
@@ -1,37 +1,22 @@
-import {
-  Logger,
-  Player,
-  Project,
-  ProjectMetadata,
-  Stage,
-} from '@motion-canvas/core';
 import type {
+  Project,
+  Player as PlayerType,
+  Stage as StageType,
   FullSceneDescription,
   ThreadGeneratorFactory,
-} from '@motion-canvas/core/lib/scenes';
-import type {View2D} from '@motion-canvas/2d/lib/components';
-import {makeScene2D} from '@motion-canvas/2d';
-import {ValueDispatcher} from '@motion-canvas/core/lib/events';
-import runtime from '@site/src/components/Fiddle/runtime';
-import DefaultPlugin from '@motion-canvas/core/lib/plugin/DefaultPlugin';
-
-declare global {
-  interface Window {
-    mc: typeof runtime & {
-      makeScene2D: unknown;
-    };
-  }
-}
+} from '@motion-canvas/core';
+import type {View2D} from '@motion-canvas/2d';
 
 let ProjectInstance: Project = null;
 let Description: FullSceneDescription<ThreadGeneratorFactory<View2D>> = null;
-let PlayerInstance: Player = null;
-let StageInstance: Stage = null;
-let CurrentSetter: (value: Player) => void = null;
+let PlayerInstance: PlayerType = null;
+let StageInstance: StageType = null;
+let CurrentSetter: (value: PlayerType) => void = null;
 let CurrentParent: HTMLElement = null;
 let CurrentRatio = 1;
+let ErrorHandler: (message: string) => void = null;
 
-export function disposePlayer(setter: (value: Player) => void) {
+export function disposePlayer(setter: (value: PlayerType) => void) {
   if (CurrentSetter !== setter || !ProjectInstance) return;
   PlayerInstance.deactivate();
   CurrentSetter = null;
@@ -39,22 +24,43 @@ export function disposePlayer(setter: (value: Player) => void) {
   StageInstance.finalBuffer.remove();
 }
 
-export function borrowPlayer(
-  setter: (value: Player) => void,
+export function updatePlayer(description: typeof Description) {
+  if (Description) {
+    Description.onReplaced.current = description;
+  }
+}
+
+export async function borrowPlayer(
+  setter: (value: PlayerType) => void,
   parent: HTMLDivElement,
   ratio: number,
-): Player {
+  onError?: (message: string) => void,
+): Promise<PlayerType> {
   if (setter === CurrentSetter) return;
+  if (
+    StageInstance &&
+    CurrentParent &&
+    StageInstance.finalBuffer.parentElement === CurrentParent
+  ) {
+    CurrentParent?.removeChild(StageInstance.finalBuffer);
+  }
+  CurrentSetter?.(null);
+  CurrentSetter = setter;
+  CurrentParent = parent;
+  ErrorHandler = onError;
 
   if (!ProjectInstance) {
-    window.mc = {
-      ...runtime,
-      makeScene2D: config => {
-        Description.config = config;
-        Description.onReplaced.current = Description;
-        return Description;
-      },
-    };
+    const {
+      Logger,
+      Player,
+      ProjectMetadata,
+      Stage,
+      ValueDispatcher,
+      DefaultPlugin,
+    } = await import(/* webpackIgnore: true */ '@motion-canvas/core');
+    const {makeScene2D} = await import(
+      /* webpackIgnore: true */ '@motion-canvas/2d'
+    );
 
     Description = makeScene2D(function* () {
       yield;
@@ -90,14 +96,13 @@ export function borrowPlayer(
       }
     });
 
-    // TODO Find a way to attach the logger only for DEV:
-    // ProjectInstance.logger.onLogged.subscribe(console.log);
+    ProjectInstance.logger.onLogged.subscribe(payload => {
+      if (payload.level === 'error') {
+        ErrorHandler?.(`Runtime error: ${payload.message}`);
+      }
+    });
   }
 
-  CurrentParent?.removeChild(StageInstance.finalBuffer);
-  CurrentSetter?.(null);
-  CurrentSetter = setter;
-  CurrentParent = parent;
   if (CurrentRatio !== ratio) {
     ProjectInstance.meta.shared.size.set([960, Math.floor(960 / ratio)]);
     Description.onReplaced.current = {
@@ -115,10 +120,15 @@ export function borrowPlayer(
   return PlayerInstance;
 }
 
-export function tryBorrowPlayer(
-  setter: (value: Player) => void,
+export async function tryBorrowPlayer(
+  setter: (value: PlayerType) => void,
   parent: HTMLDivElement,
   ratio: number,
-): Player | null {
-  return CurrentSetter ? null : borrowPlayer(setter, parent, ratio);
+  onError?: (error: string) => void,
+): Promise<PlayerType | null> {
+  if (!CurrentSetter) {
+    return borrowPlayer(setter, parent, ratio, onError);
+  }
+
+  return null;
 }