+++ packages/ui/src/components/sidebar/index.ts
@@ -1,3 +1,4 @@
 export * from './Properties';
+export * from './Settings';
 export * from './Threads';
 export * from './VideoSettings';