+++ packages/ui/src/components/viewport/EditorPreview.tsx
@@ -9,12 +9,13 @@ import {
   useSize,
   useStorage,
   useSubscribable,
+  useSubscribableValue,
 } from '../../hooks';
 import {Debug} from './Debug';
 import {Grid} from './Grid';
 import styles from './Viewport.module.scss';
 import {ViewportContext, ViewportState} from './ViewportContext';
-import {isInspectable} from '@motion-canvas/core/lib/scenes/Inspectable';
+import {isInspectable} from '@motion-canvas/core';
 import {useApplication, useInspection} from '../../contexts';
 import {highlight} from '../animations';
 import {PreviewStage} from './PreviewStage';
@@ -23,11 +24,15 @@ import {Button, Select} from '../controls';
 import {Recenter, Grid as GridIcon} from '../icons';
 import {ButtonCheckbox} from '../controls/ButtonCheckbox';
 import {ColorPicker} from './ColorPicker';
+import {Coordinates} from './Coordinates';
 
 const ZOOM_SPEED = 0.1;
 
 export function EditorPreview() {
-  const {player} = useApplication();
+  const {player, settings: appSettings} = useApplication();
+  const coordinateSetting = useSubscribableValue(
+    appSettings.appearance.coordinates.onChanged,
+  );
   const scene = useCurrentScene();
   const {setInspectedElement} = useInspection();
   const containerRef = useRef<HTMLDivElement>();
@@ -49,6 +54,7 @@ export function EditorPreview() {
   const state: ViewportState = useMemo(() => {
     const state = {
       grid,
+      size,
       width: size.width,
       height: size.height,
       ...position,
@@ -246,6 +252,7 @@ export function EditorPreview() {
             <GridIcon />
           </ButtonCheckbox>
           <ColorPicker />
+          {coordinateSetting && <Coordinates />}
         </div>
       </div>
     </ViewportContext.Provider>